backend/requirements.txt
# ContentAgent Backend Dependencies

# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# HTTP Client
requests>=2.31.0

# LLM & AI
openai>=1.57.0
httpx>=0.28.0

# Vector Database (Qdrant)
qdrant-client>=1.12.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
loguru>=0.7.0

# File Parsing
PyMuPDF>=1.24.0  # For PDF parsing
python-docx>=0.8.11  # For DOCX parsing
python-multipart>=0.0.6  # For file uploads
